+++ src/Repository.php
@@ -55,7 +55,7 @@ final class Repository
                     ->attempt(static fn($error) => new \RuntimeException($error::class)),
             )
             ->map(static fn() => new self(
-                new Binary(
+                Binary::of(
                     $server,
                     $path,
                     $home,
@@ -71,10 +71,7 @@ final class Repository
     public function init(): Attempt
     {
         return ($this->binary)(
-            $this
-                ->binary
-                ->command()
-                ->withArgument('init'),
+            static fn($command) => $command->withArgument('init'),
         )
             ->map(
                 static fn($output) => $output
@@ -87,7 +84,7 @@ final class Repository
                     false => Attempt::error(new \RuntimeException($output->toString())),
                 },
             )
-            ->map(static fn() => new SideEffect);
+            ->map(SideEffect::identity(...));
     }
 
     /**
@@ -97,9 +94,7 @@ final class Repository
     public function head(): Attempt
     {
         return ($this->binary)(
-            $this
-                ->binary
-                ->command()
+            static fn($command) => $command
                 ->withArgument('branch')
                 ->withOption('no-color'),
         )
@@ -120,7 +115,7 @@ final class Repository
     #[\NoDiscard]
     public function branches(): Branches
     {
-        return new Branches($this->binary);
+        return Branches::of($this->binary);
     }
 
     /**
@@ -130,11 +125,8 @@ final class Repository
     public function push(): Attempt
     {
         return ($this->binary)(
-            $this
-                ->binary
-                ->command()
-                ->withArgument('push'),
-        )->map(static fn() => new SideEffect);
+            static fn($command) => $command->withArgument('push'),
+        )->map(SideEffect::identity(...));
     }
 
     /**
@@ -144,29 +136,26 @@ final class Repository
     public function pull(): Attempt
     {
         return ($this->binary)(
-            $this
-                ->binary
-                ->command()
-                ->withArgument('pull'),
-        )->map(static fn() => new SideEffect);
+            static fn($command) => $command->withArgument('pull'),
+        )->map(SideEffect::identity(...));
     }
 
     #[\NoDiscard]
     public function remotes(): Remotes
     {
-        return new Remotes($this->binary);
+        return Remotes::of($this->binary);
     }
 
     #[\NoDiscard]
     public function checkout(): Checkout
     {
-        return new Checkout($this->binary);
+        return Checkout::of($this->binary);
     }
 
     #[\NoDiscard]
     public function tags(): Tags
     {
-        return new Tags($this->binary, $this->clock);
+        return Tags::of($this->binary, $this->clock);
     }
 
     /**
@@ -176,12 +165,10 @@ final class Repository
     public function add(Path $file): Attempt
     {
         return ($this->binary)(
-            $this
-                ->binary
-                ->command()
+            static fn($command) => $command
                 ->withArgument('add')
                 ->withArgument($file->toString()),
-        )->map(static fn() => new SideEffect);
+        )->map(SideEffect::identity(...));
     }
 
     /**
@@ -191,13 +178,11 @@ final class Repository
     public function commit(Message $message): Attempt
     {
         return ($this->binary)(
-            $this
-                ->binary
-                ->command()
+            static fn($command) => $command
                 ->withArgument('commit')
                 ->withShortOption('m')
                 ->withArgument($message->toString()),
-        )->map(static fn() => new SideEffect);
+        )->map(SideEffect::identity(...));
     }
 
     /**
@@ -207,12 +192,10 @@ final class Repository
     public function merge(Branch $branch): Attempt
     {
         return ($this->binary)(
-            $this
-                ->binary
-                ->command()
+            static fn($command) => $command
                 ->withArgument('merge')
                 ->withArgument($branch->toString()),
-        )->map(static fn() => new SideEffect);
+        )->map(SideEffect::identity(...));
     }
 
     /**