+++ src/Repository/Branches.php
@@ -8,6 +8,7 @@ use Innmind\Git\{
     Revision\Branch,
     Revision\Hash,
 };
+use Innmind\Server\Control\Server\Command;
 use Innmind\Immutable\{
     Set,
     Str,
@@ -18,8 +19,16 @@ use Innmind\Immutable\{
 
 final class Branches
 {
-    public function __construct(private Binary $binary)
+    private function __construct(private Binary $binary)
     {
+    }
+
+    /**
+     * @internal
+     */
+    public static function of(Binary $binary): self
+    {
+        return new self($binary);
     }
 
     /**
@@ -29,9 +38,7 @@ final class Branches
     public function local(): Set
     {
         return ($this->binary)(
-            $this
-                ->binary
-                ->command()
+            static fn($command) => $command
                 ->withArgument('branch')
                 ->withOption('no-color'),
         )
@@ -58,9 +65,7 @@ final class Branches
     public function remote(): Set
     {
         return ($this->binary)(
-            $this
-                ->binary
-                ->command()
+            static fn($command) => $command
                 ->withArgument('branch')
                 ->withShortOption('r')
                 ->withOption('no-color'),
@@ -98,17 +103,16 @@ final class Branches
     #[\NoDiscard]
     public function new(Branch $name, Hash|Branch|null $off = null): Attempt
     {
-        $command = $this
-            ->binary
-            ->command()
+        $map = static fn(Command $command): Command => $command
             ->withArgument('branch')
             ->withArgument($name->toString());
 
         if ($off) {
-            $command = $command->withArgument($off->toString());
+            $map = static fn(Command $command): Command => $map($command)
+                ->withArgument($off->toString());
         }
 
-        return ($this->binary)($command)->map(static fn() => new SideEffect);
+        return ($this->binary)($map)->map(SideEffect::identity(...));
     }
 
     /**
@@ -117,14 +121,12 @@ final class Branches
     #[\NoDiscard]
     public function newOrphan(Branch $name): Attempt
     {
-        $command = $this
-            ->binary
-            ->command()
-            ->withArgument('checkout')
-            ->withOption('orphan')
-            ->withArgument($name->toString());
-
-        return ($this->binary)($command)->map(static fn() => new SideEffect);
+        return ($this->binary)(
+            static fn($command) => $command
+                ->withArgument('checkout')
+                ->withOption('orphan')
+                ->withArgument($name->toString()),
+        )->map(SideEffect::identity(...));
     }
 
     /**
@@ -134,13 +136,11 @@ final class Branches
     public function delete(Branch $name): Attempt
     {
         return ($this->binary)(
-            $this
-                ->binary
-                ->command()
+            static fn($command) => $command
                 ->withArgument('branch')
                 ->withShortOption('d')
                 ->withArgument($name->toString()),
-        )->map(static fn() => new SideEffect);
+        )->map(SideEffect::identity(...));
     }
 
     /**
@@ -150,12 +150,10 @@ final class Branches
     public function forceDelete(Branch $name): Attempt
     {
         return ($this->binary)(
-            $this
-                ->binary
-                ->command()
+            static fn($command) => $command
                 ->withArgument('branch')
                 ->withShortOption('D')
                 ->withArgument($name->toString()),
-        )->map(static fn() => new SideEffect);
+        )->map(SideEffect::identity(...));
     }
 }