+++ tests/BinaryTest.php
@@ -4,10 +4,7 @@ declare(strict_types = 1);
 namespace Tests\Innmind\Git;
 
 use Innmind\Git\Binary;
-use Innmind\Server\Control\{
-    Server\Command,
-    Servers\Mock,
-};
+use Innmind\Server\Control\Servers\Mock;
 use Innmind\Url\Path;
 use Innmind\BlackBox\PHPUnit\Framework\TestCase;
 
@@ -31,21 +28,18 @@ class BinaryTest extends TestCase
                 );
             });
 
-        $bin = new Binary(
+        $bin = Binary::of(
             $server,
             Path::of('/tmp/foo'),
         );
 
-        $this->assertInstanceOf(Command::class, $bin->command());
-        $this->assertSame('git', $bin->command()->toString());
-        $this->assertSame('/tmp/foo', $bin->command()->workingDirectory()->match(
-            static fn($path) => $path->toString(),
-            static fn() => null,
-        ));
-        $this->assertSame(0, $bin($bin->command()->withArgument('watev'))->match(
-            static fn($output) => $output->size(),
-            static fn() => null,
-        ));
+        $this->assertSame(
+            0,
+            $bin(static fn($command) => $command->withArgument('watev'))->match(
+                static fn($output) => $output->size(),
+                static fn() => null,
+            ),
+        );
     }
 
     public function testReturnNothingWhenCommandFailed()
@@ -68,12 +62,12 @@ class BinaryTest extends TestCase
                 static fn($_, $builder) => $builder->failed(),
             );
 
-        $bin = new Binary(
+        $bin = Binary::of(
             $server,
             Path::of('/tmp/foo'),
         );
 
-        $this->assertNull($bin($bin->command()->withArgument('watev'))->match(
+        $this->assertNull($bin(static fn($command) => $command->withArgument('watev'))->match(
             static fn($output) => $output,
             static fn() => null,
         ));
@@ -81,15 +75,41 @@ class BinaryTest extends TestCase
 
     public function testHomeIsAddedToTheCommandEnvironment()
     {
-        $bin = new Binary(
-            Mock::new($this->assert()),
+        $server = Mock::new($this->assert())
+            ->willExecute(function($command) {
+                $this->assertSame(
+                    "git 'watev'",
+                    $command->toString(),
+                );
+                $this->assertFalse($command->toBeRunInBackground());
+                $this->assertSame(
+                    '/tmp/foo',
+                    $command->workingDirectory()->match(
+                        static fn($path) => $path->toString(),
+                        static fn() => null,
+                    ),
+                );
+                $this->assertSame(
+                    '/Users/baptouuuu',
+                    $command->environment()->get('HOME')->match(
+                        static fn($home) => $home,
+                        static fn() => null,
+                    ),
+                );
+            });
+
+        $bin = Binary::of(
+            $server,
             Path::of('/tmp/foo'),
             Path::of('/Users/baptouuuu'),
         );
 
-        $this->assertSame('/Users/baptouuuu', $bin->command()->environment()->get('HOME')->match(
-            static fn($home) => $home,
-            static fn() => null,
-        ));
+        $this->assertSame(
+            0,
+            $bin(static fn($command) => $command->withArgument('watev'))->match(
+                static fn($output) => $output->size(),
+                static fn() => null,
+            ),
+        );
     }
 }